Overhaul of bootstrap script, read on for details:
@yrmt and I recently discussed the state of the Save OS X project.
It's clear that @jperkin, over at Joyent, is ahead of the game, and
ruthless with his work on pkgsrc - including signed 64bit binary packages
for Darwin.

Rather than step on his toes - or, rather heels, it seems - we figure:
seeing as Joyent have been so kind as to host their tree & binary packages
to all for free, why not embrace this and combine our efforts?

This commit moves our main repo over to Joyent's at pkgsrc.joyent.com
It'll function just as it has done before, but Jonathan had already built 2015Q1,
so; fresh packages!

Save OS X is now more of a gateway to the awesomeness of pkgsrc for Darwin users
out there. We hope with this change and our combined efforts, we'll help more and
more people discover/realise just how great it is.

Regarding the actual code: there's some clean up/minor enhancements to follow :)